move utility functions to existing packages
