+++ server/subscribe.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/mattermost/mattermost-server/v5/model"
 
+	v1 "github.com/nathanaelhoun/mattermost-plugin-circleci/server/circle/v1"
 	"github.com/nathanaelhoun/mattermost-plugin-circleci/server/store"
-	"github.com/nathanaelhoun/mattermost-plugin-circleci/server/utils"
 )
 
 const (
@@ -177,7 +177,7 @@ func executeSubscribeChannel(p *Plugin, context *model.CommandArgs, split []stri
 	// TODO add message "add the orb, here is the docs for doing it"
 	return p.sendEphemeralResponse(context, fmt.Sprintf(
 		"Successfully subscribed this channel to notifications from **%s**\nSend webhooks to `%s`",
-		utils.GetFullNameFromOwnerAndRepo(owner, repo),
+		v1.GetFullNameFromOwnerAndRepo(owner, repo),
 		p.getWebhookURL(),
 	)), nil
 }
@@ -197,7 +197,7 @@ func executeUnsubscribeChannel(p *Plugin, context *model.CommandArgs, split []st
 
 	if removed := subs.RemoveSubscription(context.ChannelId, owner, repo); !removed {
 		return p.sendEphemeralResponse(context, fmt.Sprintf("This channel is not subscribed to **%s**",
-			utils.GetFullNameFromOwnerAndRepo(owner, repo))), nil
+			v1.GetFullNameFromOwnerAndRepo(owner, repo))), nil
 	}
 
 	if err := p.Store.StoreSubscriptions(subs); err != nil {
@@ -207,7 +207,7 @@ func executeUnsubscribeChannel(p *Plugin, context *model.CommandArgs, split []st
 
 	return p.sendEphemeralResponse(context, fmt.Sprintf(
 		"Successfully unsubscribed this channel to notifications from **%s**",
-		utils.GetFullNameFromOwnerAndRepo(owner, repo),
+		v1.GetFullNameFromOwnerAndRepo(owner, repo),
 	)), nil
 }
 
@@ -237,7 +237,7 @@ func executeSubscribeListAllChannels(p *Plugin, context *model.CommandArgs, spli
 		), nil
 	}
 
-	message := "Channels of this team subscribed to **" + utils.GetFullNameFromOwnerAndRepo(owner, repo) + "**\n"
+	message := "Channels of this team subscribed to **" + v1.GetFullNameFromOwnerAndRepo(owner, repo) + "**\n"
 	for _, channelID := range channelIDs {
 		channel, appErr := p.API.GetChannel(channelID)
 		if appErr != nil {
